+++ ElectricalProgressive-Core/Utils/EnergyPacket.cs
@@ -31,7 +31,7 @@ namespace ElectricalProgressive.Utils
         /// <summary>
         /// Откуда мы пришли в каждой точке пути.
         /// </summary>
-        public readonly int[] facingFrom;
+        public readonly byte[] facingFrom;
 
         /// <summary>
         /// Какие грани каждого блока уже обработаны.
@@ -59,7 +59,7 @@ namespace ElectricalProgressive.Utils
             int Voltage,
             int CurrentIndex,
             BlockPos[] Path,
-            int[] FacingFrom,
+            byte[] FacingFrom,
             bool[][] NowProcessedFaces,
             Facing[] UsedConnections
         )
